backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0 